listen to port 19990-19996, remove debug message box
Previously it was listening to 19990-19999, but we can only have 6 players in a multiplayer GTA2 game (this is enough for one display anyway!).
When limiting the possible ports to 6 instead of 10, the connecting code works faster, because it iterates through all of them sequentially.
Of course, these ports must not be used by other applications at the same time.