add frame around input bytes, that get sent via TCP
As already written in the injected_api.h file, the proxydll will be
expanded to have a whole set of API calls in the future. Therefore a
frame that indicates the used API-Call is necessary (it is just one byte
with the ID!).
Also the "input_server" files and functions got renamed to
"injected_thread", because the new code will be able to do much more
than handle input :)