rewrite debug log function ia_server_log()
It uses the FRAMEDATASEND function now and doesn't produce garbage
anymore. Also the print_hex parameter is gone (to get hex output, just
send the new IA_GARBAGE byte).

Oh and there was an annoying bug in ia_server.c (FRAMEDATASEND macro):
When the brackets are wrong in the 'NAME##_t data=' line, '+1' means
plus the size of the NAME##_t struct instead of one byte (!!!)