sty_parser: read SPRX chunk
Moved the generated offset table to the wiki:
https://github.com/Bytewerk/gta2-hackers-remix/wiki/Offsets-in-STY-files

I've added the sprite_index_t, which is not in the documentation, to
make the whole thing more consistent with font_base_t. We need to save
the sprite count somewhere, anyway.
Also I've fixed the FONB code and made it consistent with the SPRX code.