Start work on injected component
Thanks to Jester01 from freenode, it is possible now to compile without
the MSVC compiler. He helped me port the proxy assembly code to GNU
Assembly, so it is possible to use mingw now.

This is so great, now everything can be compiled on Linux \o/ (soon with
one "make" command)!

I was writing C code instead of C++ anyway, so I've "ported" the main
function to C99 now, and I'll port the rest, too :)